+++ src/scatrans/__init__.py
@@ -69,7 +69,7 @@ __all__ = [
 try:
     from ._version import version as __version__
 except ImportError:
-    __version__ = "0.9.0"
+    __version__ = "0.9.2"
 
 # Optional: prevent some internal modules from appearing too prominently
 # in casual inspection while still allowing advanced users to do